#54 moved cooperation state into index.ts moved action and mutation types into types.ts updated documentation
