+++ src/shims-vuex.d.ts
@@ -1,8 +1,9 @@
 import { Store } from '@/store';
+import { RootStateInterface } from '@/store/types';
 
 declare module '@vue/runtime-core' {
   // provide typings for `this.$store`
   interface ComponentCustomProperties {
-    $store: Store;
+    $store: Store<RootStateInterface>;
   }
 }
