+++ src/store/cooperation/getters.ts
@@ -1,11 +1,7 @@
 import { GetterTree } from 'vuex';
 
 import { RootStateInterface } from '@/store/types';
-import { CooperationInterface } from '@/store/cooperation/types';
-
-export type Getters<S = CooperationInterface> = {
-  getHelloWorldTwice(state: S): string;
-};
+import { CooperationInterface, Getters } from '@/store/cooperation/types';
 
 export const getters: GetterTree<CooperationInterface, RootStateInterface> & Getters = {
   getHelloWorldTwice: (state) => {
