--- src/store/cooperation/state.ts
@@ -1,5 +1,0 @@
-import { CooperationInterface } from '@/store/cooperation/types';
-
-export const state: CooperationInterface = {
-  helloWorld: 'hello world',
-};
