+++ src/store/cooperation/types.ts
@@ -1,9 +1,12 @@
-// названия для мутаций - a.k.a mutation enums
+import { ActionContext } from 'vuex';
+import { RootStateInterface } from '@/store/types';
+
+/** mutation enums - названия для мутаций */
 export enum CooperationMutationTypes {
   SAY_HELLO = 'SAY_HELLO',
 }
 
-// названия для экшенов
+/** action enums - названия для экшенов */
 export enum CooperationActionTypes {
   SAY_HELLO = 'SAY_HELLO',
 }
@@ -16,3 +19,43 @@ export enum CooperationActionTypes {
 export interface CooperationInterface {
   helloWorld: string;
 }
+
+/**
+ * mutation contracts
+ * типы мутаций - список мутаций их принимаемые аргументы и возвращаемое значение
+ */
+export type Mutations<S = CooperationInterface> = {
+  [CooperationMutationTypes.SAY_HELLO](state: S, payload: string): void;
+};
+
+/** action contracts */
+export interface Actions {
+  [CooperationActionTypes.SAY_HELLO]({ commit }: AugmentedActionContext): void;
+}
+
+/** getter contracts */
+export type Getters<S = CooperationInterface> = {
+  getHelloWorldTwice(state: S): string;
+};
+
+/**
+ * action context, переопределение
+ * этот тип нужен для того, чтобы разрешить выполнение только тех экшенов, которые определены в Mutations
+ * также здесь проверяется тип payload
+ * компилятор подсветит, если будет вызван/передан неверный экшен/payload
+ * */
+
+export type AugmentedActionContext = {
+  commit<K extends keyof Mutations>(key: K, payload: Parameters<Mutations[K]>[1]): ReturnType<Mutations[K]>;
+} & Omit<ActionContext<CooperationInterface, RootStateInterface>, 'commit'>;
+
+/* eslint-disable */
+/**
+ * links
+ * почитать:
+ * - https://dev.to/3vilarthas/vuex-typescript-m4j
+ * - https://dev.to/shubhadip/vue-3-vuex-4-modules-typescript-2i2o
+ * посмотреть (vue 2) - https://www.youtube.com/watch?v=wRK1EZpicD0&t
+ * потыкать - https://codesandbox.io/s/using-vuex-4-modules-in-vue-3-with-typescript-7ygvy?file=/src/store/modules/documents/mutations.ts
+ * проинвестигейтить - https://github.com/shubhadip/vuex-typescript
+ * */
